+++ src/Indice.Features.Cases.App/src/app/features/cases/cases.component.ts
@@ -68,7 +68,7 @@ export class CasesComponent extends BaseListComponent<CasePartial> implements On
                 multiTerm: true
             }
             for (let checkpointType of checkpointTypes) { // fill checkpointTypeSearchOption's SelectInputOptions
-                checkpointTypeSearchOption.options?.push({ value: checkpointType, label: checkpointType })
+                checkpointTypeSearchOption.options?.push({ value: checkpointType?.code, label: checkpointType?.title ?? checkpointType?.code! })
             }
             this.searchOptions = [
                 {