+++ src/Indice.Features.Cases.AspNetCore/Controllers/AdminCheckpointTypesController.cs
@@ -1,5 +1,6 @@
 ﻿using System.Net.Mime;
 using Indice.Features.Cases.Interfaces;
+using Indice.Features.Cases.Models.Responses;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -24,13 +25,11 @@ internal class AdminCheckpointTypesController : ControllerBase
         _checkpointTypeService = checkpointTypeService ?? throw new ArgumentNullException(nameof(checkpointTypeService));
     }
 
-    /// <summary>Get Checkpoint types</summary>
+    /// <summary>Get the distinct checkpoint types grouped by code</summary>
     [HttpGet]
     [Produces(MediaTypeNames.Application.Json)]
-    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(List<string>))]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(List<CheckpointType>))]
     [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(ProblemDetails))]
-    public async Task<IActionResult> GetDistinctCheckpointCodes() {
-        var checkpointTypes = await _checkpointTypeService.GetDistinctCheckpointCodes(User);
-        return Ok(checkpointTypes);
-    }
+    public async Task<IActionResult> GetDistinctCheckpointCodes() =>
+        Ok(await _checkpointTypeService.GetDistinctCheckpointTypes(User));
 }