+++ src/Indice.Features.Cases.AspNetCore/Models/Responses/CheckpointType.cs
@@ -1,4 +1,6 @@
-﻿using Indice.Types;
+﻿using Indice.Features.Cases.Data.Models;
+using Indice.Types;
+using Newtonsoft.Json.Converters;
 
 namespace Indice.Features.Cases.Models.Responses;
 
@@ -16,6 +18,13 @@ public class CheckpointType
 
     /// <summary>The description of the checkpoint.</summary>
     public string Description { get; set; }
+
+    /// <summary>The status of the case. This is the external status for the customer.</summary>
+    [Newtonsoft.Json.JsonConverter(typeof(StringEnumConverter))] // Because of Elsa.HttpActivities the default serializer is newtonsoft
+    public CaseStatus? Status { get; set; }
+
+    /// <summary>Indicates if the checkpoint is private, which means not visible to the customer.</summary>
+    public bool? Private { get; set; }
 
     /// <summary>The translations of the checkpoint.</summary>
     public TranslationDictionary<CheckpointTypeTranslation> Translations { get; set; }