+++ src/Indice.Features.Cases.AspNetCore/Services/AdminCaseService.cs
@@ -170,6 +170,7 @@ internal class AdminCaseService : BaseCaseService, IAdminCaseService
             throw new ResourceUnauthorizedException();
         }
         @case.CaseType = @case.CaseType.Translate(CultureInfo.CurrentCulture.TwoLetterISOLanguageName, true);
+        @case.CheckpointType = @case.CheckpointType.Translate(CultureInfo.CurrentCulture.TwoLetterISOLanguageName, true);
         return @case;
     }
 
@@ -295,15 +296,25 @@ internal class AdminCaseService : BaseCaseService, IAdminCaseService
                 CreatedBy = c.CreatedBy,
                 Checkpoint = new Checkpoint {
                     Id = c.Id,
-                    Private = c.CheckpointType.Private,
-                    CheckpointTypeCode = c.CheckpointType.Code,
+                    CheckpointType = new CheckpointType {
+                        Id = c.CheckpointType.Id,
+                        Code = c.CheckpointType.Code,
+                        Description = c.CheckpointType.Description,
+                        Translations = TranslationDictionary<CheckpointTypeTranslation>.FromJson(c.CheckpointType.Translations),
+                        Private = c.CheckpointType.Private,
+                        Status = c.CheckpointType.Status
+                    },
                     CompletedDate = c.CompletedDate,
-                    DueDate = c.DueDate,
-                    Status = c.CheckpointType.Status
+                    DueDate = c.DueDate
                 }
             }))
             .OrderByDescending(c => c.Timestamp)
-            .ThenBy(c => c.IsCheckpoint);
+            .ThenBy(c => c.IsCheckpoint)
+            .ToList();
+
+        foreach (var timelineEntry in timeline.Where(x => x.Checkpoint is not null)) {
+            timelineEntry.Checkpoint.CheckpointType = timelineEntry.Checkpoint.CheckpointType.Translate(CultureInfo.CurrentCulture.TwoLetterISOLanguageName, true);
+        }
 
         return timeline;
     }