+++ src/Indice.Features.Cases.AspNetCore/Services/CheckpointTypeService.cs
@@ -1,7 +1,11 @@
-﻿using System.Security.Claims;
+﻿using System.Globalization;
+using System.Security.Claims;
 using Indice.Features.Cases.Data;
+using Indice.Features.Cases.Data.Models;
 using Indice.Features.Cases.Interfaces;
+using Indice.Features.Cases.Models.Responses;
 using Indice.Security;
+using Indice.Types;
 using Microsoft.EntityFrameworkCore;
 
 namespace Indice.Features.Cases.Services;
@@ -14,15 +18,15 @@ internal class CheckpointTypeService : ICheckpointTypeService
         _dbContext = dbContext ?? throw new ArgumentNullException(nameof(dbContext));
     }
 
-    public async Task<List<string>> GetDistinctCheckpointCodes(ClaimsPrincipal user) {
+    public async Task<IEnumerable<CheckpointType>> GetDistinctCheckpointTypes(ClaimsPrincipal user) {
         if (user.IsAdmin()) {
-            return await GetAdminDistinctCheckpointNames();
+            return await GetAdminDistinctCheckpoints();
         }
 
         var roleClaims = user.Claims
-           .Where(c => c.Type == BasicClaimTypes.Role)
-           .Select(c => c.Value)
-           .ToList();
+            .Where(c => c.Type == BasicClaimTypes.Role)
+            .Select(c => c.Value)
+            .ToList();
 
         var checkpointTypeIds = await _dbContext.Members
             .AsQueryable()
@@ -30,22 +34,43 @@ internal class CheckpointTypeService : ICheckpointTypeService
             .Select(c => c.CheckpointTypeId)
             .ToListAsync();
 
-        var checkpointTypes = await _dbContext.CheckpointTypes
-            .AsQueryable()
-            .Where(c => checkpointTypeIds.Contains(c.Id))
-            .Select(c => c.Code)
-            .AsAsyncEnumerable()
-            .Distinct() // TODO client-side evaluation, this needs to change
-            .ToListAsync();
-        return checkpointTypes;
+        var checkpointTypes = await (
+                from c in _dbContext.CheckpointTypes
+                where checkpointTypeIds.Contains(c.Id)
+                group c by new { c.Code, c.Title } into grouped
+                select grouped.FirstOrDefault()
+            ).ToListAsync();
+
+        return TranslateCheckpoints(checkpointTypes);
     }
 
-    private async Task<List<string>> GetAdminDistinctCheckpointNames() {
-        return await _dbContext.CheckpointTypes
-            .AsQueryable()
-            .Select(c => c.Code)
-            .AsAsyncEnumerable()
-            .Distinct() // TODO client-side evaluation, this needs to change
-            .ToListAsync();
+    private async Task<IEnumerable<CheckpointType>> GetAdminDistinctCheckpoints() {
+        var checkpointTypes = await (
+            from c in _dbContext.CheckpointTypes
+            group c by new { c.Code, c.Title } into grouped
+            select grouped.FirstOrDefault()
+        ).ToListAsync();
+
+        return TranslateCheckpoints(checkpointTypes);
+    }
+
+    private static IEnumerable<CheckpointType> TranslateCheckpoints(IEnumerable<DbCheckpointType> checkpointTypes) {
+        var translated = checkpointTypes
+            .Select(c => new CheckpointType {
+                Id = c.Id,
+                Code = c.Code,
+                Title = c.Title,
+                Description = c.Description,
+                Translations = TranslationDictionary<CheckpointTypeTranslation>.FromJson(c.Translations)
+            })
+            .ToList();
+
+        foreach (var item in translated) {
+            var translation = item.Translate(CultureInfo.CurrentCulture.TwoLetterISOLanguageName, false);
+            item.Title = translation.Title ?? item.Title;
+            item.Description = translation.Description ?? item.Description;
+        }
+
+        return translated;
     }
 }