+++ src/Indice.Features.Cases.UI/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `CheckpointType` Translations to cases list, checkpointtype filter, case details and timeline.
+
 ## [7.1.8] - 2023-05-26
 ## Bugfix
 - CaseForm now allows empty layout and shows json schema correctly